+++ network/peer/config.go
@@ -4,6 +4,7 @@
 package peer
 
 import (
+	"sync/atomic"
 	"time"
 
 	"github.com/ava-labs/avalanchego/ids"
@@ -58,4 +59,8 @@ type Config struct {
 
 	// Signs my IP so I can send my signed IP address in the Handshake message
 	IPSigner *IPSigner
+
+	// IngressConnectionCount counts the ingress (to us) connections.
+	// Needs to be a pointer because it's shared across all peer connections.
+	IngressConnectionCount *atomic.Uint32
 }